refactor into matrix utils
