+++ cellxgene_schema_cli/cellxgene_schema/utils.py
@@ -8,17 +8,11 @@ from typing import Dict, List, Union
 import anndata as ad
 import h5py
 import numpy as np
-from anndata.compat import DaskArray
 from anndata.experimental import read_dispatched, read_elem_as_dask
 from cellxgene_ontology_guide.ontology_parser import OntologyParser
-from dask.array import map_blocks
-from scipy import sparse
 from xxhash import xxh3_64_intdigest
 
 logger = logging.getLogger(__name__)
-
-SPARSE_MATRIX_TYPES = {"csr", "csc", "coo"}
-SUPPORTED_SPARSE_MATRIX_TYPES = {"csr"}
 
 KB = 1024
 MB = 1024 * KB
@@ -75,31 +69,6 @@ def remap_deprecated_features(*, adata: ad.AnnData, remapped_features: Dict[str,
         raw_adata.var.index = [remapped_features.get(val, val) for val in raw_adata.var.index]
         adata.raw = raw_adata
     return adata
-
-
-def get_matrix_format(matrix: DaskArray) -> str:
-    """
-    Given a matrix, returns the format as one of: csc, csr, coo, dense
-    or unknown.
-
-    This mimics the scipy.sparse `format` property, but extends it to
-    support ndarray and other classes AnnData may proxy the matrix with.
-    """
-
-    # Note: the AnnData proxy classes DO support the `format_str` property, but
-    # doing a slice seemed safer, if less performant.  Using `format_str`, which
-    # currently works, uses private API:
-    #
-    # >>> return getattr(matrix, "format_str", "dense)
-    #
-    matrix_format = "unknown"
-    matrix_slice = matrix[0:1, 0:1].compute()
-    if isinstance(matrix_slice, sparse.spmatrix):
-        matrix_format = matrix_slice.format
-    elif isinstance(matrix_slice, np.ndarray):
-        matrix_format = "dense"
-    assert matrix_format in SPARSE_MATRIX_TYPES.union({"unknown", "dense"})
-    return matrix_format
 
 
 def getattr_anndata(adata: ad.AnnData, attr: str = None):
@@ -198,48 +167,3 @@ def is_ontological_descendant_of(onto: OntologyParser, term: str, target: str, i
 @lru_cache()
 def get_descendants(onto: OntologyParser, term: str, include_self: bool = True) -> List[str]:
     return onto.get_term_descendants(term, include_self=True)
-
-
-def count_matrix_nonzero(matrix: DaskArray, is_sparse_matrix: bool) -> int:
-    def count_nonzeros(matrix_chunk: Union[np.ndarray, sparse.spmatrix], is_sparse_matrix: bool) -> np.array:
-        nnz = matrix_chunk.nnz if is_sparse_matrix else np.count_nonzero(matrix_chunk)
-        return np.array([nnz])
-
-    if len(matrix.chunks[0]) > 1:
-        nonzeros = map_blocks(count_nonzeros, matrix, is_sparse_matrix, drop_axis=1, dtype=int).compute().sum()
-    else:
-        nonzeros = count_nonzeros(matrix.compute(), is_sparse_matrix)[0]
-    return nonzeros
-
-
-def check_non_csr_matrices(adata: ad.AnnData):
-    """
-    Check X, raw.X and layers matrices for having more than 50% zeros and not being csr_matrix
-
-    If found, convert to csr_matrix
-    """
-
-    def get_sparsity(matrix: DaskArray, format: str):
-        is_sparse_matrix = format in SPARSE_MATRIX_TYPES
-        nnz = count_matrix_nonzero(matrix, is_sparse_matrix)
-        sparsity = 1 - nnz / np.prod(matrix.shape)
-        return sparsity
-
-    format = get_matrix_format(adata.X)
-    if format != "csr" and get_sparsity(adata.X, format) >= 0.5:
-        adata.X = adata.X.map_blocks(sparse.csr_matrix, dtype=adata.X.dtype)
-
-    if adata.raw is not None:
-        format = get_matrix_format(adata.raw.X)
-        if format != "csr" and get_sparsity(adata.raw.X, format) >= 0.5:
-            raw_adata = ad.AnnData(adata.raw.X, var=adata.raw.var, obs=adata.obs)
-            raw_adata.X = raw_adata.X.map_blocks(sparse.csr_matrix, dtype=raw_adata.X.dtype)
-            adata.raw = raw_adata
-            del raw_adata
-
-    for layer in adata.layers:
-        format = get_matrix_format(adata.layers[layer])
-        if format != "csr" and get_sparsity(adata.layers[layer], format) >= 0.5:
-            adata.layers[layer] = adata.layers[layer].map_blocks(sparse.csr_matrix, dtype=adata.layers[layer].X.dtype)
-
-    return adata
